backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-jose
passlib[bcrypt]
bcrypt
python-multipart
pydantic
google-genai
requests
python-dotenv
pillow
numpy
opencv-python
scikit-learn
torch
torchvision
pypdf